+++ marvel_app/lib/src/shared/widget/single_item_card_widget.dart
@@ -1,25 +1,20 @@
-
 import 'package:flutter/material.dart';
 
 class SingleItemCard extends StatelessWidget {
   final String? label;
   final void Function()? onPressed;
-  final ImageProvider backgroundImageUrl;
+  final ImageProvider backgroundImage;
   final EdgeInsets padding;
   final EdgeInsets margin;
+
   const SingleItemCard({
     super.key,
-     this.label,
+    this.label,
     this.onPressed,
-   required this.backgroundImageUrl,
+    required this.backgroundImage,
     this.padding = const EdgeInsets.all(8.0),
-    this.margin = const EdgeInsets.symmetric(
-      vertical: 16.0,
-      horizontal: 21.0,
-    ),
+    this.margin = const EdgeInsets.symmetric(horizontal: 30, vertical: 24),
   });
-
- 
 
   @override
   Widget build(BuildContext context) {
@@ -30,16 +25,12 @@ class SingleItemCard extends StatelessWidget {
         child: Container(
           decoration: BoxDecoration(
             image: DecorationImage(
-              image: 
-               backgroundImageUrl,
-            
+              image: backgroundImage,
               fit: BoxFit.cover,
             ),
             border: Border.all(
-              color: Colors.red,
-              width: 2,
+              width: 3,
             ),
-            borderRadius: BorderRadius.circular(8.0),
           ),
           child: Padding(
             padding: margin,
